+++ ExilenceNextApp/src/components/item-table/item-table-filter/ItemTableFilter.styles.ts
@@ -3,7 +3,10 @@ import { makeStyles } from '@material-ui/core';
 const useStyles = makeStyles(theme => ({
   searchField: {
     width: '100%'
-  }
+  },
+  clearIcon: {
+    marginRight: theme.spacing(-0.5)
+  },
 }));
 
 export default useStyles;