+++ ExilenceNextApp/src/components/item-table/item-table-filter/ItemTableFilter.tsx
@@ -1,4 +1,4 @@
-import { TextField } from '@material-ui/core';
+import { TextField, IconButton } from '@material-ui/core';
 import { Formik } from 'formik';
 import { observer } from 'mobx-react';
 import React, { ChangeEvent } from 'react';
@@ -6,17 +6,20 @@ import { useTranslation } from 'react-i18next';
 import * as Yup from 'yup';
 import { IPricedItem } from '../../../interfaces/priced-item.interface';
 import useStyles from './ItemTableFilter.styles';
+import CloseIcon from '@material-ui/icons/Close';
 
 export interface TableFilterProps<T> {
   array: T[];
   handleFilter: (
     event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
   ) => void;
+  clearFilter: () => void;
 }
 
 const ItemTableFilter: React.FC<TableFilterProps<IPricedItem>> = ({
   array,
-  handleFilter
+  handleFilter,
+  clearFilter
 }: TableFilterProps<IPricedItem>) => {
   const { t } = useTranslation();
   const classes = useStyles();
@@ -33,7 +36,7 @@ const ItemTableFilter: React.FC<TableFilterProps<IPricedItem>> = ({
     >
       {/* todo: refactor and use new formik */}
       {(formProps: any) => {
-        const { values, handleChange } = formProps;
+        const { values, setFieldValue, handleChange } = formProps;
 
         return (
           <form>
@@ -44,9 +47,31 @@ const ItemTableFilter: React.FC<TableFilterProps<IPricedItem>> = ({
                 handleChange(e);
                 handleFilter(e);
               }}
-              id="search-text"
+              name="searchText"
               label={t('tables:label.search_text')}
               className={classes.searchField}
+              value={values.searchText}
+              InputProps={{
+                endAdornment: (
+                  <>
+                    {values.searchText !== '' && (
+                      <IconButton
+                        aria-label="help"
+                        title={t('label.clear_search_icon_title')}
+                        className={classes.clearIcon}
+                        edge="start"
+                        size="small"
+                        onClick={e => {
+                          setFieldValue('searchText', '');
+                          clearFilter();
+                        }}
+                      >
+                        <CloseIcon />
+                      </IconButton>
+                    )}
+                  </>
+                )
+              }}
             />
           </form>
         );