+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.3.2] - NOT RELEASED
+##### Added
+- Added support for pricing delirium orbs
+- Added support for pricing vials
 ##### Fixed
 - Fixed a bug where the overlay sometimes would not stay on top of the game